Report max 50 file comments to github
The github api limits us to report 50 file comments per api
request. Since the check run reporting is in the critical path don't
do paging but limit this to 50 comments. In order to get a diverse
feedback sort the messages by 'uniqueness' first so that 100 equal
comments don't push out 5 other comments that are valuable.

Change-Id: If5a7c246a58b510d8fcfa31bd71d74c84acbc278