Update patch set 6
Patch Set 6:

It looks like in response to my previous comment, you have just taken that code and moved it into a separate project maintained by the SF team.

I'm not sure what the problem with react-ansi is -- does it not meet our needs in some way?  It seems like using an existing project with a wider contributor base is desirable as long as it works for us.

Patch-set: 6