Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

I agree that making the entire row clickable had unintended consequences.  However, I think this change presents significant usability problems.  A basic tenant of human-computer interaction is the concept of "affordances".  For example, the knob on a stove is an indication that the temperature of the stove can be adjusted.  Here we are hiding the most important action that a user can perform on this page -- the hidden magnifying glass means there is no visible affordance to navigate off of the page.

Links are the main way of affording the option of navigating on the WWW, and they are usually styled in such a way to indicate that.

The most obvious places for a user to expect a link to a result page for an individual build are either the result column "It says 'failure', I will click on 'failure' to see the failure" or the job name "I will click on 'tox-linters' to see the result of that job".  Let's make one or both of those links to the result page.

Patch-set: 4
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1