Update patch set 6
Patch Set 6:

Has anyone evaluated react-ansi or re-ansi with very large (10,000 line) files?  ISTR some concern about ansi libs handling that in our earlier attempts.  It seems like that should be a significant criterion for deciding what lib to use (and I think it was omitted in the analysis on the react-ansi change).

Patch-set: 6