Update patch set 4
Patch Set 4:

I get the point that this is more like an experimental feature which might not be accepted by all. However, the intention of this change was to bring both worlds together. I know a lot of people who found it very unintuitive in the old UI that you have to click on the result to go to the build result page - me included. Although it might sound reasonable to say "I will click on 'failure' to see the failure" - IIRC the failure never looked like a clickable link in any way. So the idea behind this change was to put a dedicated icon that serves the link. Maybe hiding it by default wasn't a good idea, although I think that a user usually hovers over the table with the mouse anyway and thus should get noticed by the "pop-up" icon that there is something you can act on.

Long story short: Thinking about this now - with the new UI where also the job_name and the result are colored (in contrary to the rest of the columns) - it might feel a little more correct to put a link on both, so I will give that a try.

Patch-set: 4
Reviewer: Gerrit User 27952 <27952@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0