Increase default zookeeper session timeout
The default session timeout currently is 10s. On larger zuul
deployments even a full gc run can cause a longer stall of the zk
heartbeat thread and lead to a session loss which has a severe impact
on the system. By default zookeeper server allows for 40s max session
timeout so pick that one as a default.

Change-Id: I7f3d6744454d534aeb701d2027fa0c4bb7ce274f