web: return 0 as severity when line doesn't match
Since Ibcbc2bd9497f1d8b75acd9e4979a289173d014b2 we have inadvertently
dropped lines that don't match any regex.  This is because their
severity is set to null, but in the log viewer page we decide to show
a line based on the match

     (line.severity >= severity)

so "null >= <integer>" returns false and we don't see those lines.

By ensuring lines that don't match have a severity of 0 we filter
these appropriately.

Change-Id: Ie402db32b723ddfe447aa2bb16ccaf369e070e57