web: Add optional link prop to title-with-icon and results
This adds an optional link prop to the BuildResultWithIcon and
BuildResult components which is a ... link :) These are used in other
contexts (like titles) where it doesn't make sense to have them as
links, which is why I kept it optional.

Additionally this is inside the span and inherits the color; it is
intentional that these links retain their underline (that was the
original review request).

The external Link wrappers used in various places are removed now.

Change-Id: Ib4c3a2159e7bad58fcc27eb3ba4b62926ce73099