Update patch set 2
Patch Set 2: Code-Review-1

The compressed UUID display is a unique UI pattern and I initially had no idea the whole thing was selected and was trying to click the ... to make it expand.  These values are of little use and I am in favour of removing these, or otherwise hiding them to reclaim the header space for the timing (which is what I thought was proposed in the discussion), but I'm -1 on doing it like this.

As a suggestion, if we feel the need to keep them in the header I think something like

 <icon> UUID >

and you click on that and it pulls down to reveal the values would make sense; and by default only take up as much space as this anyway.

However, personally I think they're available quickly from the inventory.yaml and would be in favour of just removing them completely to simplify things.

To my eye this has become unnecessarily compressed.  I'd want something to convince me that this is better than accepting the PF4 default css.  I'm not -1 on that, but personally I don't find it compelling without some more justification.

Patch-set: 2
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1