Update patch set 4
Abandoned

Abandoning as it seems that there is no enough traction to make this happen.

The only accepted outcome was that it was o to add "yarn lint" to tox -e linters, so we have a single command for linting all code.

I will be happy to restore the change if the situation changes.

Patch-set: 4
Status: abandoned