Handle the yaml merge operator
The << operator in yaml is currently handled as a key and not like an
operator. This leads to a YAMLDuplicateKeyError when parsing yaml that
merges multiple mappings into the same dict.

This can be solved by not saving this operator into our key list and
let pyyaml handle the rest.

Change-Id: Ic444747b9dc6fd64396fdb9944faa4c41c2bcd31