Use logical timestamp to detect outdated changes
In a multi-scheduler deployment the scheduler processing a connection
event is not necessarily the same one that is later processing the
global or pipeline trigger event. To only refresh a change in case it is
outdated we can use the Zookeeper 'zxid' as a logical timestamp for the
change and attach it also to the event.

Change-Id: I6a0e58c1fb6d3e3ee798e8a76d85e7b3ae61f77b