Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> Since we have agreement this is not worth taking up header space for, I think we should simply remove it

Removal is outside the scope and it was not discussed or part of the proposal. I am personally
against removal of build-id (i never used the event id) and the benefit would be zero if only 
one goes away.

Also, I would prefer to make changes progressively and allow others to come with proposals. Maybe we can find a better place to host this information.

The main reason for progressive approach is that I want users to slowly learn the bits and avoid being hit by too much. This also allows us to tune the implementation based on the feedback.

Patch-set: 5
Reviewer: Gerrit User 24162 <24162@4a232e18-c5a9-48ee-94c0-e04e7cca6543>