Update patch set 3
Patch Set 3:

Thanks! Looking more closely after our IRC discussion, I think this actually points out some deeper problems with our terminology... We have a "child_jobs" variable here which is about (among other things) the order in which jobs are run:

https://zuul-ci.org/docs/zuul/reference/jobs.html#var-zuul.child_jobs

But then we have the "parent" attribute in job definitions:

https://zuul-ci.org/docs/zuul/reference/job_def.html#attr-job.parent

A bit of keyword searching further indicates that we then use "parent job" in a number of places to also refer to jobs which run before other jobs, and we refer to "child job" in some places when we're talking about a job which inherits content from another job definition. While I hate to say it, I think we're actually going to need to rename something and deprecate whatever we rename to properly untangle this clash in terminology, so this is going to prompt a much larger discussion.

Patch-set: 3
CC: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>