Cache unparsed config files in Zookeeper
Cache config files in Zookeeper so we can load the config on multiple
schedulers with only one round trip to the mergers.

In contrast to how it's described in the scale out scheduler spec, the
cache will be namespaced by tenant. This has the advantage that we can
do a read/write lock of the layout on a per tenant level which doesn't
block reconfigurations of other tenants. The downside of this approch is
that we might have some duplicate files in Zookeeper in case a project
is part of multiple tenants.

Change-Id: Ia3d2a124eaf6ee9d82c94a0375ee7ecba61caadf