Update patch set 34
Patch Set 34: Code-Review-1

It looks like if a user suplies multiple sql connections we're only going to use the one that says default=true, which make the others redundant.  And since we require the user to add "default=true" to the config file, that means a config file change is required in every instance where there is more than one sql connection.

So I think before we merge this, we should come to a decision about whether we want to support more than one connection.  If we only want to support one, then we should make the switch to a [database] section and remove the default= parameter (but still suppport the current syntax for smooth upgrades).  If we want to support more than one connection, then we should not add the [database] section, but should go ahead and add tenant sql connection parameter (and keep the default parameter to use in case no tenant sql connection is supplied).

Patch-set: 34
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1