Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I wonder if we should first add a deprecation message to the script and keep it for a while.
> 
> My main concern is that this code is untested, whereas zuul-client's encrypt command is validated in functional tests. I'd like also to avoid changes to this script that may not be backported to zuul-client.

Another option would be to replace the script content to only write a message that this has been replaced by zuul-client together with a link to the zuul-client docs. I'd like to avoid that our end users (who don't follow the mailing list or release notes) are surprised by a non-existing script and having to ask our ops team what to use instead.

Patch-set: 2