Deprecates old admin client
Documents the deprecation of the admin client and points user
to the newer zuul-client alternative. This should help users searching
online for information about the client, to reach the newer tool.

Once we are confident that the old admin client can be removed we
will also remove it from the docs.

Change-Id: I643a879ae235fc17981f30c7770b0f51fe4c80da