Fix zuul-client enqueue-ref when oldrev/newrev aren't provided
According to zuul-client documentation [1] and source code [2], using
enqueue-ref without oldrev/newrev is a valid use case. In such case,
oldrev and newrev are both set to null reference [2]

Adding a test creates the following exception:

  File "/home/testzuul/zuul/zuul/.tox/py38/lib/python3.8/site-packages/git/refs/symbolic.py", line 302, in set_reference
    obj = self.repo.rev_parse(ref + "^{}")    # optionally deref tags
  File "/home/testzuul/zuul/zuul/.tox/py38/lib/python3.8/site-packages/git/repo/fun.py", line 213, in rev_parse
    obj = name_to_object(repo, rev[:start])
  File "/home/testzuul/zuul/zuul/.tox/py38/lib/python3.8/site-packages/git/repo/fun.py", line 147, in name_to_object
    raise BadName(name)
gitdb.exc.BadName: Ref 'master' did not resolve to an object

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "/home/testzuul/zuul/zuul/zuul/executor/server.py", line 943, in execute
    self._execute()
  File "/home/testzuul/zuul/zuul/zuul/executor/server.py", line 1121, in _execute
    repo.checkout(selected_ref)
  File "/home/testzuul/zuul/zuul/zuul/merger/merger.py", line 526, in checkout
    repo.head.reference = ref
  File "/home/testzuul/zuul/zuul/.tox/py38/lib/python3.8/site-packages/git/refs/symbolic.py", line 305, in set_reference
    raise ValueError("Could not extract object from %s" % ref) from e
ValueError: Could not extract object from master

This is due to the ref being cleaned during getRepoState because newrev
is equal to a null reference.
Merger is set to avoid this when oldrev and newrev are both null.

[1] https://zuul-ci.org/docs/zuul-client/commands.html#enqueue-ref
[2] https://opendev.org/zuul/zuul-client/src/branch/master/zuulclient/cmd/__init__.py#L106-L109

Change-Id: I0f91194e4a76af87d68b1a402456119412323c5e