Store tenant layout state in Zookeeper
To coordinate reconfigurations across schedulers we need to store the
current layout state in Zookeeper. We store the hostname of the
scheduler that processed the reconfiguration event as well as the human
readable timestamp of the reconfig.

The ltime of the layout state is the last modified transaction ID of the
corresponding znode in Zookeeper. It will later be used to detect layout
changes on other schedulers.

Change-Id: I243aa4f54b038fbad78e331fbeeb508989e34b5f