Document tox environments
Adds descriptions tox environments in order to make it easier
to identify which tests should be run.

This should also update at least one of the two doc pages below in
order to mention tox and most important commands.

https://zuul-ci.org/docs/zuul/reference/developer/index.html
https://zuul-ci.org/docs/zuul/reference/developer/testing.html

Also correct the py3 and py3-docker expansion list to correctly
use a comma.

Change-Id: Iddfb9919ab66f1b661966719b812d598e3cb0c91