Enable git protocol v2 in merger
Git versions between 2.18 and 2.26 have version 2
protocol support but it's not enabled by default.
Starting in 2.26, it is the default.  V2 is more
efficient at negotiating refs and can reduce the time
spent updating a repo by 50% on large repos.  Enable
it in the merger code for newly cloned repos.

Change-Id: Ifc471c5fde49c65137bc34e49734b3ac3d33e3e1