Don't bail out if change behind cannot be queried
In case there is a draft change behind a change that is to be enqueued
it can be that zuul has no access to it. In this case zuul currently
bails out completely during building the git-needed by or
commit-needed by lists that are not strictly necessary for the change
in question. Catch exceptions in this case and ignore those changes
behind.

Change-Id: I995c69f27668d172a0eddfdf6547266334f215c7