Support cross scheduler config loading
On startup a scheduler will read/generate the local tenant config and
start loading the configuration. For this it checks if there is already
a layout state in Zookeeper for a given tenant:

- In case a layout state already exists the scheduler willl acquire the
  tenant READ lock, read the cached unparsed config from Zookeeper and
  load the tenant.
- If there is no existing layout state for a tenant the scheduler
  acquires the tenant WRITE lock, loads the tenant and creates a new
  layout state in Zookeeper.

After a tenant reconfiguration the layout state in Zookeeper for that
tenant is update. All other schedulers are notified of the changed
layout state via a data watch and will reload the changed tenant and
update the local layout state.

As long as the local layout state of a scheduler doesn't match the
tenant's layout state in Zookeeper it is not allowed to process any
pipelines for that tenant.

Change-Id: I48b145cecec5890d212d751f68b3947f04ee1ed5