[WIP] web UI: user login with OpenID Connect
*** Set as WIP to investigate other oidc libraries ***

If the default realm's authentication driver is set to "OpenIDConnect",
display a "Sign in" button. If the the user is logged in, fetch user
authorizations and add them to the redux store; display the user's
preferred username in the upper right corner. Clicking on the user icon
in the right corner displays a modal with user information such as
the user's zuul-client configuration, and a sign out button.

Clicking on the sign out button removes user information from the
store (note that it does not log the user out from the identity provider).

Add some basic documentation explaining how to configure Zuul with
Google's authentication, and with a Keycloak server.

Authentication is intended to be tenant-scoped, but at this time the
REST API only provides one auth configuration for all tenants.
A follow-up patch allows linking a specific realm (and thus
authentication) to a tenant.

Change-Id: I31e71f2795f3f7c4253d0d5b8ed309bfd7d4f98e