Update patch set 5
Patch Set 5:

I agree that this looks like a good change to do if we want to set cors from all API paths (which is what we currently do).

However, I wonder if we made a mistake there and we should set this only for read-only paths (such as the status API, which is what I think we started setting cors headers for in the first place).  Should the new writable APIs (such as enqueue/dequeue, etc) really have CORS headers?

Patch-set: 5
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>