Example Docker compose: keycloak integration
Add a keycloak container to the compose to handle users in gerrit and
zuul services.

Due to the way containers work, using keycloak requires to map localhost
to "keycloak" in /etc/hosts when running this compose. This is only
required when working with zuul or gerrit web UIs; the REST APIs are
unaffected.

This patch will always be rebased on the tip of the topic branch, so
that all changes can be tested at once. In order to test a specific
change, it is possible to cherry pick this patch on top of that change.

Below is the list of the configuration steps that are automated with the
compose:

Keycloak:
* Create keycloak admin user (admin:kcadmin) that can be used to further
  configure the service, for example to add social identity providers.
  The admin console can be reached at http://keycloak:8082
* Provision the "zuul-demo" realm with users admin, user1, user2 (passwords
  are their respective usernames). Create "tenant-one-admin" group, to which
  user1 belongs.
* Create and configure OpenID Connect clients for zuul and gerrit. The
  zuul OIDC client is set so that it will allow authentication requests
  from a development UI (provided it runs on the default port 3000).
Gerrit:
* Configure OpenID Connect authentication with keycloak
* Pre-provision admin account with pynotedb. Set admin's http password
  to secret.
Zuul Web:
* Configure keycloak authenticator
Zuul Scheduler:
* Add admin rules on example-tenant, so that admin and user1 can use the
  admin REST API.

It is up to the tester to provision gerrit with a project and to
configure it properly in zuul (it must at least belong to example-tenant
for the predefined admin rules to apply).

Change-Id: Id6dc22f73e323ea65bfc546a9971b190495b95c5