Example Docker compose: keycloak integration
Add a keycloak container to the compose to handle users in gerrit and
zuul services.

Due to the way containers work, using keycloak requires to map 127.0.0.1
to "keycloak" in /etc/hosts when running this compose. This is only
required when working with zuul or gerrit web UIs; the REST APIs are
unaffected.

Keycloak:
* Create keycloak admin user (admin:kcadmin) that can be used to further
  configure the service, for example to add social identity providers.
  The admin console can be reached at http://keycloak:8082
* Provision zuul-realm with users admin, user1, user2 (passwords are their
  respective usernames). Create "tenant-one-admin" group, to which user1
  belongs.
* Create and configure OpenID Connect clients for zuul and gerrit.
Gerrit:
* Configure OpenID Connect authentication with keycloak
* Pre-provision admin account with pynotedb. Set admin's http password
  to secret.
Zuul Web:
* Configure keycloak authenticator
Zuul Scheduler:
* Add admin rules on example-tenant, so that admin and user1 can use the
  admin REST API.

Change-Id: Id6dc22f73e323ea65bfc546a9971b190495b95c5