Call r.text only once
When querying the gerrit api we call the text property of the result
object several times. Accessing this property can be expensive since
it does text conversion and eventually charset guessing each
time. Further this property is not cached within the response object
so query the text only once and reuse it. This can considerably
improve performance on a busy zuul using gerrit.

Change-Id: I4ad4eba979b36fa07e68ecc6c0ef4039a4440bf6