Implement ZooKeeper builds API
This allows us to store builds in ZooKeeper. The builds are stored with
their uuid and namespaced by their zone, resulting in a path like:

    /zuul/builds/<zone>/<build_uuid>/[cancel, resume]

Each build node stored in ZooKeeper might contain a subnode "cancel" or
"resume" which indicates one of those actions to be executed by the
executor server on this build.

Due to the complex data model in Zuul, we decided to create a new
BuildItem class which is serializable and contains a minimal set of data
which is necessary to be stored in ZooKeeper.
Thus, in the current state we will have a Build and a BuildItem for each
build running in Zuul. Having both classes is only an intermediate step
until the Build class itself is fully serializable and can be stored in
ZooKeeper.

Change-Id: I4ed89c608cad21b8aa2895b2090831f0db17892d