Switch to ZooKeeper backed job execution and result events
Builds and their lifecycle events (BuildStarted, BuildStatus,
BuildPaused, BuildCompleted) will be dispatched via ZooKeeper.

This change still relies on local Build objects being present on the
scheduler in addition to the BuildItems being stored in ZooKeeper.

The executor server uses a TreeCacheListener to listen for changes on
the build and actively picks up new builds in its main loop.

This change moves most of the result event logic in either the scheduler
or the executor server class. Thus, the executor server updates the
BuildItem on ZooKeeper accordingly with each lifecycle transition and
puts a BuildResultEvent in ZooKeeper. Those events are then picked up by
the scheduler which updates the local Build object accordingly and
notifies the pipeline managers like before.

In case a build is finished, the scheduler returns the node set and
removes the BuildItem in ZooKeeper.

Co-authored-by: Jan Kubovy <jan.kubovy@bmw.de>
Change-Id: I5de26afdf6774944b35472e2054b93d12fe21793