Update patch set 45
Patch Set 45:

There are a bunch of comments from Clark that have not been responded to here.  I see that the linked change does address some of them, but it's important that we have a coherent development history.  We should *at least* reply to these comments with answers -- something like "yes, I changed this in ..." or "no, we don't need to because..." is fine.  But even better would be doing that and then squashing the followup change into this one.

Patch-set: 45
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>