gitlab: Add access token name, Update docs, Fix webhook
Highlight that Access Token:
- can be created at project level and user level
- can be used to authenticate when cloning repositories

To clone internal or private projects, authentication is required. It
can be achieved using the Access Token. Currently, any names can be used
but [1] shows that it's likely going to be restricted to username and
token-name. As token-name is directly related to an Access Token, use
this value.

webhook can be created foreach project but also at group level, so at
the topmost group. Therefor only one webhook at the upper group can be
used to send events to Zuul: remove "project" from webhook_token
description

[1] # https://gitlab.com/gitlab-org/gitlab/-/issues/212953

Change-Id: I57adb5e03658d31bd3900d2d9037101cf491466e