Implement ZooKeeper backed merge jobs
This implements the base functionality to execute merge jobs with
ZooKeeper instead of Gearman as backend an prepares the merger and
executor servers to use the new implementation.

The merge jobs will be stored in
    /zuul/merges/<uuid>

The old Gearman approach is still working with this change, the
transition from one to another will be done next.

Change-Id: I6762a535b147adab9e1192f028112aec7997ac07