Store semaphore state in Zookeeper
In order for semaphores to work in a multi-scheduler environment we need
to store their state in Zookeeper.

The state for acquired semaphores is stored in
`/zuul/semaphores/<tenant>/<semaphore>/<queue-item-uuid>-<job-name>`. In
order to acquire or release a semaphore a scheduler must hold the lock
for a semaphore. Those are stored in
`/zuul/semaphore_locks/<tenant>/<semaphore>`.

Change-Id: Iddc71ae26008cea081838b1001d33a98ed05c0d8