Store semaphore state in Zookeeper
In order for semaphores to work in a multi-scheduler environment we need
to store their state in Zookeeper.

The state for acquired semaphores is stored in
`/zuul/semaphores/<tenant>/<semaphore>`. The semaphore node will hold a
list of handles in the form `<queue-item-uuid>-<job-name>`.

The semaphore handler will try to acquire/release a semaphore
optimistically and retry in case of concurrent updates.

Change-Id: Iddc71ae26008cea081838b1001d33a98ed05c0d8