Perform per tenant locking in getStatus
When retrieving the status json zuul-web currently holds a global
lock. This serializes all status requests regardless of the need to
update. This can lead to very slow response times in a loaded multi
tenant environment.

This can be improved by locking per tenant and also don't block if we
can satisfy the request by the cache.

Change-Id: I9896bedd79761c304066027b1735a517511692ce