Switch to ZooKeeper backed merge jobs
Merge jobs jobs will now be dispatched via ZooKeeper.

Instead of notifying the client first, the merger server now puts the
merge result directly in the corresponding result queue in ZooKeeper. In
case the caller wants to wait for the result synchronously (e.g. during
config loading), the merger will provide a MergeResultEventFuture
instead which contains the path to the result in ZooKeeper.

Change-Id: Icd2e29f154505c8b17c52db06890ca1b1cd097c3