Format all log entries on one line
There are two significant cases in Zuul where we may emit multi-line
log messages:

* Tracebacks
* Zuul config errors

When these are sent to a log file or stderr using the default
configurations, it can be difficult to work with those logs.  Using
"grep" to search for messages or time periods will typically only
produce a single line of the error.

Further, on a busy system the error may be interleaved with other
log lines making extraction even harder.

To address both of these issues, add a formatter which concats every
line of a multi-line log message together with the "|" character.
This does make it harder to read the exceptions in the simple case
(a slow system without grep, etc).  But that is outweighed by the
fact that it makes it possible to obtain all the data in the more
complex case.  Exceptions and error messages are still somewhat
readable even with this formatting.  In extreme cases (say an
exceptionally long traceback) someone could always reverse the process
on the exception of interest.

Since this is implemented as a standard logging formatter, it can
be opted-out via logging config files.  This enables it by default,
which I think is reasonable for new users, but if we decide that we
would prefer opt-in, that is a simple change.

Change-Id: I6d7e7e7a9e19d46a744f9ffac8d532fc6b4bba01