Fix test_gerrit.TestPolling.test_config_update
Currently, this test triggers multiple tenant reconfigurations because
the change search query in the FakeGerritConnection does not evaluate
the age parameter. So far, this wasn't a problem because the events
where processed directly and the test could settle fast enough.

With events dispatched via ZooKeeper, there will always be an event in
any of the queues, so the test cannot settle.

Change-Id: I54f7c9bb340bd9b1c74a363077c6da79b5480e69