Provide zk_client to merger client rather than the whole scheduler
With the transition from Gearman to ZooKeeper for merge jobs, the merger
client doesn't need to notify the scheduler anymore, as the merger
server directly puts the merge results in the corresponding result event
queue.

Thus, instead of the scheduler instance, we only provide the ZooKeeper
client to the merger client.

Change-Id: Ia61701e7c82153e955703dd564d414a0a44dc84c