Remove history from RecordingMergeClient
There are quite a lot of test cases which assert that certain merge jobs
where run by taking a look at the gearman_server.job_history. With the
transition to ZooKeeper this was replaced by using the
merge_job_queue.history instead.

However, there are two test cases which use another history that is
stored on the RecordingMergeClient. To reduce the number of "history
approaches" and align those test cases with the rest, let's switch to
the merge_job_queue.history instead for those tests.

The RecordingMergeClient is now only used to overwrite the
_merge_job_queue_class for the tests. If we find another way to do so,
we could remove the RecordingMergeClient completely.

Change-Id: I46da5d23abbf69879cc1da18219b58f83472af4d