Remove history from RecordingMergeClient
There are quite a lot of test cases which assert that certain merge jobs
where run by taking a look at the gearman_server.job_history. With the
transition to ZooKeeper this was replaced by using the
ZuulTestCase.merge_job_history property instead.

However, there are two test cases which use another history that is
stored on the RecordingMergeClient. To reduce the number of "history
approaches" and align those test cases with the rest, let's switch to
the merge_job_history property instead for those tests.

This also renames the RecordingMergeClient to HoldableMergeClient as
this now doesn't record anything anymore (and to align with the name of
the HoldableExecutorApi class that is used in the tests).

Change-Id: I46da5d23abbf69879cc1da18219b58f83472af4d