Add waiting cause to status API
When a job has not made a node request, this adds a text description
of the reason to the status JSON.  This currently appears as "waiting"
on the status page, but users (including this author) report being
confused as to what the job is waiting for and how that differs from
the "queued" state.

With this change in place, we will be able to update the status page
to report something like "waiting for dependencies: other-job-name"
in mouseover text over the waiting status.

Change-Id: I73b0a8b8c21507826ccc32d3f7c580b6c853ca6d