Make scheduler optional in Nodepool
To lock/unlock the nodes directly in the executor server, we have to
make the Nodepool API work without a scheduler instance.

This also makes the stats emitting in the nodepool class optional as
this relies on the scheduler being provided. This is just an
intermediate step as we will have to find a different solution for the
stats reporting once multiple schedulers are in place.

The calls to self.sched.onNodesProvisioned() won't be necessary anymore
once the executor directly locks/unlocks the nodes.

Change-Id: I3f3e4bfff08e244f68a9be7c6a4efcc194a23332