Update patch set 19
Patch Set 19:

(1 comment)

I've tried this on our deployment and it worked out of the box 😊. However I've one question inline and there is one thing I think that needs to be fixed before merge. On page reload zuul-web forgets that the user is already logged in which is ok for a poc but confuses users.

Patch-set: 19
CC: Gerrit User 16068 <16068@4a232e18-c5a9-48ee-94c0-e04e7cca6543>