Update patch set 9
Patch Set 9: Code-Review+1

Not a bad idea, I had to do the same on other projects in order to reuse test code in related projects.

The cost of including the tests should not really matter much. In addition to this I do also
see this as a possible way to use some of them to verify a deployment.

Patch-set: 9
Reviewer: Gerrit User 24162 <24162@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1