Update patch set 9
Patch Set 9: Code-Review-2

We explicitly chose not to make the tests part of the Zuul deliverable.  We're not going to guarantee an external-facing API, and due to all the test fixtures, they are *huge*.  I think we need a different approach to solve this problem.

Patch-set: 9
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2